+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.12
 - Convert OpenCore-Patcher binary to OpenCore-Patcher.app
 - Add Backlight patches for modded Nvidia GPUs in iMac10,x-12,x
+- Fix sleep for iMac12,x with upgraded GPUs
 
 ## 0.0.11
 - Re-add OpenCore GUI